requirements.txt

pyserial
pyyaml
watchdog
